% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.exp_miRNA.R
\docType{data}
\name{case.exp_miRNA}
\alias{case.exp_miRNA}
\title{miRNA-seq expression data in the ECSA cancer samples}
\usage{
data("case.exp_miRNA")
}
\description{
miRNA-seq expression data in the ECSA cancer samples, the row names are miRNA mirbase IDs, the column names are sample names in which last two identifier is 01
}
\examples{
data("case.exp_miRNA")
head(case.exp_miRNA[1:100,])
}
