% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.exp_rna.R
\docType{data}
\name{case.exp_rna}
\alias{case.exp_rna}
\title{RNA-seq expression data in the ECSA cancer samples}
\usage{
data("case.exp_rna")
}
\description{
RNA-seq expression data in the ECSA cancer samples, the row names are RNA mirbase IDs, the column names are sample names in which last two identifier is 01
}
\examples{
data("case.exp_rna")
head(case.exp_rna[1:100,])
}
