% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cepair.r
\name{cepair}
\alias{cepair}
\title{extract lncRNA-circRNA ceRNA pair}
\usage{
cepair(interac, N_mi = 50)
}
\arguments{
\item{interac}{the interaction data matrix, in which the column name contain "node_gene_ID", "type" and  "target_gene_ID"}

\item{N_mi}{a numeric value, shared miRNAs by a lncRNA and circRNA}
}
\value{
interac_temp the interaction data matrix, in which the column name contain "node_gene_ID", "type" and  "target_gene_ID"
}
\description{
extract lncRNA-circRNA ceRNA pair
}
\examples{
\dontrun{
interac <- interStringency(type = "ncRNA", stringency = "strict")
interac_strict_p_50 <- cepair(interac, N_mi= 50)
}
}
