% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinData.R
\docType{data}
\name{clinData}
\alias{clinData}
\title{clinical data in the ECSA cancer patients}
\usage{
data("clinData")
}
\description{
clinical data in the ECSA cancer patients, the column names are TCGA sample ID(cancer sample ID),survival time (month),survival status
}
\examples{
data("clinData")
head(clinData[1:100,])
}
