% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinp.R
\name{combinp}
\alias{combinp}
\title{Generate corrected p-value based on p-value and logFC in the expression matrix}
\usage{
combinp(node_attr = NULL, islog = T)
}
\arguments{
\item{node_attr}{A data frame containing three columns: type, logFC and p value, and the row name is the gene identifier.}

\item{islog}{Boolean value, whether to use the logFC, if FALSE, the weight is the p-value, or "TRUE", the corrected p-value is used}
}
\value{
A data matrix containing three columns: type, gene, weight(corrected p-value), the row name is the gene identifier.
}
\description{
based on the formula: corrected p-value = 2*( 1-pnorm( (-log10( p-value )) * abs( log2FC )) ), corrected p-value was generated
}
\examples{
data("dataN")
result <- combinp(dataN[,c("type","logFC","PValue")])

}
\references{
Hongbo Shi, Jiayao Li, Qiong Song et al. (2019) Systematic identification and analysis of dysregulated miRNA and transcription factor feed-forward loops in hypertrophic cardiomyopathy
}
