% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.exp_miRNA.R
\docType{data}
\name{control.exp_miRNA}
\alias{control.exp_miRNA}
\title{miRNA-seq expression data in the ECSA normal samples}
\usage{
data("control.exp_miRNA")
}
\description{
miRNA-seq expression data in the ECSA normal samples, the row names are miRNA mirbase IDs, the column names are sample names in which last two identifier is 11
}
\examples{
data("control.exp_miRNA")
head(control.exp_miRNA[1:100,])
}
