% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.exp_rna.R
\docType{data}
\name{control.exp_rna}
\alias{control.exp_rna}
\title{RNA-seq expression data in the ECSA normal samples}
\usage{
data("control.exp_rna")
}
\description{
RNA-seq expression data in the ECSA normal samples, the row names are RNA mirbase IDS, the column names are sample names in which last two sign is 11
}
\examples{
data("control.exp_rna")
head(control.exp_rna[1:100,])
}
