% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataN.R
\docType{data}
\name{dataN}
\alias{dataN}
\title{Example node_attr for ncRNA network}
\usage{
data("dataN")
}
\description{
Data example consisting of a matrix of the colname of genes, type, logFC, logCPM, PValue, FDR.
}
\examples{
data("dataN")
head(dataN[1:100,])
}
