% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGs.R
\name{DEGs}
\alias{DEGs}
\title{Performe differential expression analysis}
\usage{
DEGs(case.exp, control.exp, geneid, data_type)
}
\arguments{
\item{case.exp}{the case expression matrix, in which the row name is gene id and the column name is sample id}

\item{control.exp}{the control expression matrix, in which the row name is gene id (the same with the case) and the column name is sample id}

\item{geneid}{gene id in the case or control expression matrix}

\item{data_type}{a character string indicating which date type to deal with is to be choosed, One of "RNAseq_counts" , "fpkm" and "microarray": can be abbreviated}
}
\value{
DEGlist contain four elements,
 DEGs the differential expression matrix,
 Nor_expr the normalized expression matrix,
 data_type a character string of the data type
 algorithm  a character string indicating which algorithm  was used, One of "edgeR" , "limma"
}
\description{
Differential expression analysis using edgeR or limma for two group comparison
}
\examples{
\dontrun{
data("case.exp_miRNA")
data("control.exp_miRNA")
result_miR <- DEGs(case.exp_miRNA,control.exp_miRNA,
geneid= rownames(control.exp_miRNA), data_type = "RNAseq_counts")

}
}
