% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESCA_clinical.R
\docType{data}
\name{ESCA_clinical}
\alias{ESCA_clinical}
\title{clinical data in the ECSA cancer patients}
\usage{
data(ESCA_clinical)
}
\description{
clinical data in the ECSA cancer patients, the column names are TCGA sample ID(cancer sample ID),survival time (month),survival status
}
\examples{
data(ESCA_clinical)
head(ESCA_clinical[1:100,])
}
