% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_type.R
\docType{data}
\name{gene_type}
\alias{gene_type}
\title{Table of the gene identifier and the type for ensembl ID}
\usage{
data("gene_type")
}
\description{
Data  consisting of a matrix, the colname contain gene_ID and gene_Name
}
\examples{
data("gene_type")
head(gene_type[1:100,])
}
