% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integPvals.R
\name{integPvals}
\alias{integPvals}
\title{Integrate multiple p-values into joint p-values}
\usage{
integPvals(pvalmatrix)
}
\arguments{
\item{pvalmatrix}{Numeric matrix of p-values, columns represent different sets of p-values}
}
\value{
matrix The matrix of two columns: gene, weight(p-value)
}
\description{
The function integrate multiple p-values into the joint p-value of p-values based on the order statistics of p-values.
An joint p-value #is given by the kth order statistic.
}
