% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSub.R
\name{plotSub}
\alias{plotSub}
\title{Plot of the subnetwork}
\usage{
plotSub(
  network,
  layout = layout.fruchterman.reingold,
  labels = NULL,
  diff.expr = NULL,
  scores = NULL,
  main = NULL,
  vertex.size = NULL
)
}
\arguments{
\item{network}{A graph in igraph or graphNEL format.}

\item{layout}{Layout algorithm, e.g. layout.fruchterman.reingold or layout.kamada.kawai.}

\item{labels}{Labels for the nodes of the network}

\item{diff.expr}{Named numerical vector of log2FC of the nodes in the network for coloring of the nodes.}

\item{scores}{Named numerical vector of scores of the nodes for the shape of the node in the network.}

\item{main}{Main title of the plot.}

\item{vertex.size}{Numerical value or verctor for the size of the vertices.}
}
\description{
The function plots a network from graphNEL or igraph format. It is used to visualize the modules. For further plotting options use the plot.igraph
function of the igraph package. The shapes of the nodes can be changed according to the scores argument, then negative scores appear squared
The color of the nodes can be changed according to the diff.expr argument. Negative(positive) values lead to green(red) nodes.
}
\examples{
library(igraph)
edgel <- cbind(c("1", "2", "3", "4", "5", "6", "7"),
c("b", "c", "d", "e", "f", "a", "b"))
g <- graph.edgelist(edgel, directed=TRUE)
V(g)$type <- c(rep("lncRNA",4),rep("miRNA",4),rep("circRNA",5))
plotSub(g)

}
\references{
Daniela Beisser, Gunnar W. Klau, Thomas Dandekar et al. (2010) BioNet: an R-Package for the functional analysis of biological networks
}
