% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\name{runmodule}
\alias{runmodule}
\title{Run module search function}
\usage{
runmodule(
  network,
  gene2weight,
  method = c("global", "local"),
  expr1 = NULL,
  expr2 = NULL,
  d = 2,
  r = 0.1,
  seletN = NULL,
  FDR = 1e-14,
  lambda = 0.5,
  min.size = 5,
  maxsize = 15,
  issymbol = TRUE
)
}
\arguments{
\item{network}{A data frame containing a symbolic edge list of the ncRNA network in which the columns must contain "node_gene_ID", "type", "target_gene_Name"}

\item{gene2weight}{A weigth data frame containing three columns:"type","gene", "weight"
the first "type" the type of the gene identifier; lncRNA, miRNA, circRNA and RBP
the second gene is unique, gene identifier (should be coordinate with the node symbol used in ncRNA network);
the third weight is gene-based  p-value or corrected p-value derived from differentially gene analysis or survival analysis}

\item{method}{a character string indicating which the search method is to be computed
. One of "global" (default, refer to Heinz method), "local ( refer to GS method)": can be abbreviated}

\item{d}{An integer used to define the order of neighbour genes to be
searched  in  the method of  the method "local" . This parameter is default set up as 2}

\item{r}{A float indicating the cut-off for increment during module expanding
process in  the method of  the method "local". Greater r will generate smaller module. Default is 0.1.}

\item{seletN}{a vector: gene identifier IDs, or a gene identifier ID, for example "MIMAT0000461",c("MIMAT0000461", "ENSG00000250742")}

\item{FDR}{Numeric value, from the false discovery rate a p-value threshold is calculated. P-values below this threshold are considered to be significant
The FDR can be used to control the size of the maximum scoring module}

\item{maxsize}{An integer: the numbel of size of the module for user settings in the method of "global", default 15.}

\item{issymbol}{Boolean value, whether to set the node attribute "symbol"(gene symbol) in the network.
@return \code{runmodule} returns a list containing relevant data and results,
 including:
\tabular{ll}{
  \code{GNCW} \tab the node-weighted network used for searching \cr
  \code{module} \tab list of genes comprising each module, named for the seed gene if the method is "local" or the igraph class of the module if the method is "global" \cr
  \code{module.score.matrix} \tab contains Zm, Zn \cr
}}
}
\description{
runmodule constructs a node-weighted ncRNA network, performs module searching, generates simulation data from random networks,

normalizes module scores using simulation data, removes un-qualified modules, and orders resultant modules according to their significance.
}
\examples{
\dontrun{
data("dataN")
gene2weight <- combinp(dataN[,c("type","logFC","PValue")])
interac <- interStringency(type = "transcription", spec ="hg",stringency = "strict")
interac <- interac[,c("node_gene_ID","type","target_gene_ID")]
 res.list_global <- runmodule(network = interac, gene2weight, method = "global",FDR = 1e-14)
 res.list_local <- runmodule(network = interac, gene2weight, method = "local",
 maxsize=15, seletN = "MIMAT0000461")
}

}
\references{
Hongbo Shi, Jiayao Li, Qiong Song et al. (2019) Systematic identification and analysis of dysregulated miRNA and transcription factor feed-forward loops in hypertrophic cardiomyopathy

Peilin Jia, Siyuan Zheng, Jirong Rong, Wei Zheng, Zhongming Zhao. (2011) Bioformatics. dmGWAS: dense module searching for genome-wide association studies in protein-protein interaction networks.

Daniela Beisser, Gunnar W. Klau, Thomas Dandekar  et al. (2019)  BioNet: an R-Package for the functional analysis of biological networks
}
