% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savelocalM.R
\name{savelocalM}
\alias{savelocalM}
\title{save and plot module}
\usage{
savelocalM(res.list_local)
}
\arguments{
\item{res.list_local}{the methd local result}
}
\value{
the plot and the format XGMML of the each module, filenames is the seed node
}
\description{
save and plot module for the method local result in the function runmodule
}
