% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveNetwork.R
\name{saveNetwork}
\alias{saveNetwork}
\title{save of the subnetwork}
\usage{
saveNetwork(
  network,
  name = "network",
  file,
  type = c("table", "XGMML", "sif", "tab")
)
}
\arguments{
\item{network}{A graph in igraph or graphNEL format.}

\item{name}{Name of the network, only needed for the XGMML format.}

\item{file}{File name to save.}

\item{type}{Type in which graph shall be saved.}
}
\description{
The function plots a network from graphNEL or igraph format. It is used to visualize the modules. For further plotting options use the plot.igraph
function of the igraph package. The shapes of the nodes can be changed according to the scores argument, then negative scores appear squared
The color of the nodes can be changed according to the diff.expr argument. Negative(positive) values lead to green(red) nodes.
}
\examples{
library(igraph)
edgel <- cbind(c("1", "2", "3", "4", "5", "6", "7"), c("b", "c", "d", "e", "f", "a", "b"))
g <- graph.edgelist(edgel, directed=TRUE)
V(g)$type <- c(rep("lncRNA",4),rep("miRNA",4),rep("circRNA",5))
saveNetwork(g,file ="g", type = "XGMML")
}
\references{
Daniela Beisser, Gunnar W. Klau, Thomas Dandekar et al. (2010) BioNet: an R-Package for the functional analysis of biological networks
}
