% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.km.R
\name{survival.km}
\alias{survival.km}
\title{Performe univariate Cox regression analysis and plot Kaplan-Meier curve of the RNC module}
\usage{
survival.km(
  gene_profile = NULL,
  clinData = NULL,
  genes = NULL,
  filename = "temp"
)
}
\arguments{
\item{gene_profile}{the expression value matrix, in which the row name is gene id and the column name is sample id}

\item{clinData}{the survival data, in which the column name is sample id, Survival(time) and Status(0,1)}

\item{genes}{the gene in the RNC, The name of mature miRNA is miRbase ID, and name of lncRNA and RBP is Ensemble gene ID}

\item{filename}{the output figure name in Kaplan-Meier curve}
}
\value{
contain the figure with the pdf format in Kaplan-Meier curve
}
\description{
univariate Cox regression analysis using survival with survival data and expression data,and Patients with higher and lower than the median risk score of the he 
dysregulated RNC are classified into different groups. Kaplan-Meier survival analysis was used to assess the clinical significance between the comparison groups
}
