% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.rna.R
\name{survival.rna}
\alias{survival.rna}
\title{Performe univariate Cox regression analysis}
\usage{
survival.rna(gene_profile = NULL, clinData = NULL)
}
\arguments{
\item{gene_profile}{the expression value matrix, in which the row name is gene id and the column name is sample id}

\item{clinData}{the survival data, in which the column name is sample id, Survival(time) and Status(0,1)}
}
\value{
surivallist contain two elements,
 rna_p the survival result,in which the column name is gene id, P-value(from a univariable Cox proportional
hazards regression model) and gene type(0,1)
 algorithm  a character string indicating which algorithm  was used with univariable Cox
}
\description{
univariate Cox regression analysis using survival with survival data and expression data
}
