% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interStringency.R
\name{interStringency}
\alias{interStringency}
\title{Extract interactions according to stringency and interaction type}
\usage{
interStringency(
  type = c("Protein", "transcription"),
  spec = c("hg", "mm"),
  stringency = c("low", "medium", "high", "strict")
)
}
\arguments{
\item{type}{a character string indicating which interaction type is to be choosed,
. One of "Protein" (RBP-circRNA,RBP-lncRNA,miRNA-circRNA,miRNA-lncRNA,RBP-miRNA,RBP-RBP), "transcription (miRNA-circRNA,miRNA-lncRNA,miRNA-RBP))": can be abbreviated}

\item{spec}{a character string indicating which species is to be choosed,
. One of "hg", "mm": can be abbreviated}

\item{stringency}{a character string indicating which interaction stringency is to be choosed,
. One of "low" ( number of supported experiments > = 1 or combined_score >= 150 or score >=0 ), "medium ( > = 2 or combined_score >= 400 or score >=5)","high ( > = 3 or combined_score >= 700 or score >=10)","strict ( > = 5 or combined_score >= 900, or score >=20)"}
}
\value{
interaction of setting
}
\description{
interactions were extracted according to stringency and interaction type in the database of ENCORI
}
\examples{
\dontrun{
  #data("dataM2C")
#data("dataM2L")
#data("dataM2R")
#data("dataR2C")
#data("dataR2L")
 interac <- interStringency(type = "miRNA",stringency = "strict")
 interac <- interac[,c("node_gene_ID","target_gene_ID")]
}

}
